+++ src/App.tsx
@@ -4,12 +4,13 @@ import { AnimesPage } from './pages/AnimesPage/AnimesPage'
 import { MangaPage } from './pages/MangaPage/MangaPage'
 import { PokemonPage } from './pages/PokemonPage/PokemonPage'
 import { PokemonGameDetailPage } from './pages/PokemonGameDetailPage/PokemonGameDetailPage'
+import { SkyrimPage } from './pages/SkyrimPage/SkyrimPage'
 
 /**
  * Composant racine de l'application, définissant les routes
  * principales : la page de connexion, la page des animes, la page
- * des mangas, la page des jeux Pokémon et la page de détail d'un jeu
- * Pokémon (non référencée dans la Navbar).
+ * des mangas, la page des jeux Pokémon, la page de détail d'un jeu
+ * Pokémon (non référencée dans la Navbar) et la page Skyrim.
  */
 function App() {
   return (
@@ -19,6 +20,7 @@ function App() {
       <Route path="/mangas" element={<MangaPage />} />
       <Route path="/pokemon" element={<PokemonPage />} />
       <Route path="/pokemon/:id" element={<PokemonGameDetailPage />} />
+      <Route path="/skyrim" element={<SkyrimPage />} />
     </Routes>
   )
 }